Specs moved to processors tests
