+++ src/Factory/OrderFactory.php
@@ -8,41 +8,25 @@ use Sylius\Component\Core\Model\AddressInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\Model\OrderItemInterface;
-use Sylius\Component\Core\Model\ProductVariantInterface;
 use Sylius\Component\Inventory\Checker\AvailabilityCheckerInterface;
 use Sylius\Component\Order\Modifier\OrderItemQuantityModifierInterface;
 use Sylius\Component\Order\Modifier\OrderModifierInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
+use Sylius\Component\Resource\Repository\RepositoryInterface;
+use Sylius\CustomerReorderPlugin\ReorderProcessing\ReorderProcessor;
 
 final class OrderFactory implements OrderFactoryInterface
 {
     /** @var FactoryInterface */
     private $decoratedFactory;
 
-    /** @var OrderModifierInterface */
-    private $orderModifier;
+    /** @var ReorderProcessor */
+    private $reorderProcessor;
 
-    /** @var FactoryInterface */
-    private $orderItemFactory;
-
-    /** @var OrderItemQuantityModifierInterface */
-    private $orderItemQuantityModifier;
-
-    /** @var AvailabilityCheckerInterface */
-    private $availabilityChecker;
-
-    public function __construct(
-        FactoryInterface $decoratedFactory,
-        FactoryInterface $orderItemFactory,
-        OrderModifierInterface $orderModifier,
-        OrderItemQuantityModifierInterface $orderItemQuantityModifier,
-        AvailabilityCheckerInterface $availabilityChecker
-    ) {
+    public function __construct(FactoryInterface $decoratedFactory, ReorderProcessor $reorderProcessor)
+    {
         $this->decoratedFactory = $decoratedFactory;
-        $this->orderItemFactory = $orderItemFactory;
-        $this->orderModifier = $orderModifier;
-        $this->orderItemQuantityModifier = $orderItemQuantityModifier;
-        $this->availabilityChecker = $availabilityChecker;
+        $this->reorderProcessor = $reorderProcessor;
     }
 
     public function createNew()
@@ -54,60 +38,12 @@ final class OrderFactory implements OrderFactoryInterface
 
     public function createFromExistingOrder(OrderInterface $order, ChannelInterface $channel): OrderInterface
     {
-        $reorder = $this->decoratedFactory->createNew();
+        $reorder = $this->createNew();
         assert($reorder instanceof OrderInterface);
 
         $reorder->setChannel($channel);
-//        $reorder->setCustomer($order->getCustomer());
-//        $reorder->setCurrencyCode($order->getCurrencyCode());
-//        $reorder->setNotes($order->getNotes());
-//        $reorder->setLocaleCode($order->getLocaleCode());
-//
-//        /** @var AddressInterface $billingAddress */
-//        $billingAddress = $order->getBillingAddress();
-//
-//        /** @var AddressInterface $shippingAddress */
-//        $shippingAddress = $order->getShippingAddress();
-//        $reorder->setBillingAddress(clone $billingAddress);
-//        $reorder->setShippingAddress(clone $shippingAddress);
-
-        $this->copyOrderItemsToReorder($order, $reorder);
+        $this->reorderProcessor->process($order, $reorder);
 
         return $reorder;
-    }
-
-    private function copyOrderItemsToReorder(OrderInterface $order, OrderInterface $reorder): void
-    {
-        $orderItems = $order->getItems();
-
-        /** @var OrderItemInterface $orderItem */
-        foreach ($orderItems as $orderItem) {
-            /** @var ProductVariantInterface $productVariant */
-            $productVariant = $orderItem->getVariant();
-            if ($productVariant->isTracked() && !$productVariant->isInStock()) {
-                continue;
-            }
-
-            $reorderItemQuantity = 0;
-
-            if (!$this->availabilityChecker->isStockSufficient($orderItem->getVariant(), $orderItem->getQuantity())) {
-                $reorderItemQuantity = $orderItem->getVariant()->getOnHand() - $orderItem->getVariant()->getOnHold();
-            }
-
-            else {
-                $reorderItemQuantity = $orderItem->getQuantity();
-            }
-
-            /** @var OrderItemInterface $newItem */
-            $newItem = $this->orderItemFactory->createNew();
-
-            $newItem->setVariant($orderItem->getVariant());
-            $newItem->setUnitPrice($orderItem->getUnitPrice());
-            $newItem->setProductName($orderItem->getProductName());
-            $newItem->setVariantName($orderItem->getVariantName());
-
-            $this->orderItemQuantityModifier->modify($newItem, $reorderItemQuantity);
-            $this->orderModifier->addToOrder($reorder, $newItem);
-        }
     }
 }
