+++ tests/Application/app/AppKernel.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 use Sylius\Bundle\CoreBundle\Application\Kernel;
 use Sylius\CustomerReorderPlugin\DependencyInjection\Compiler\RegisterEligibilityCheckerResponseProcessorsPass;
 use Sylius\CustomerReorderPlugin\DependencyInjection\Compiler\RegisterEligibilityCheckersPass;
+use Sylius\CustomerReorderPlugin\DependencyInjection\Compiler\RegisterReorderProcessorsPass;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
@@ -38,5 +39,6 @@ final class AppKernel extends Kernel
     {
         $container->addCompilerPass(new RegisterEligibilityCheckersPass());
         $container->addCompilerPass(new RegisterEligibilityCheckerResponseProcessorsPass());
+        $container->addCompilerPass(new RegisterReorderProcessorsPass());
     }
 }
